+++ packages/stream_chat_persistence/lib/src/db/drift_chat_database.dart
@@ -15,6 +15,7 @@ part 'drift_chat_database.g.dart';
     DraftMessages,
     Locations,
     Messages,
+    PendingOperations,
     PinnedMessages,
     Polls,
     PollVotes,
@@ -33,6 +34,7 @@ part 'drift_chat_database.g.dart';
     MessageDao,
     DraftMessageDao,
     LocationDao,
+    PendingOperationDao,
     PinnedMessageDao,
     PinnedMessageReactionDao,
     MemberDao,
@@ -58,7 +60,7 @@ class DriftChatDatabase extends _$DriftChatDatabase {
 
   // you should bump this number whenever you change or add a table definition.
   @override
-  int get schemaVersion => 1000 + 35;
+  int get schemaVersion => 1000 + 36;
 
   // Store DateTime as ISO-8601 text to preserve sub-second precision.
   @override