+++ packages/stream_chat_persistence/lib/src/stream_chat_persistence_client.dart
@@ -653,6 +653,27 @@ class StreamChatPersistenceClient extends ChatPersistenceClient {
     }
   }
 
+  @override
+  Future<int?> insertPendingOperation(PendingOperation operation) {
+    assert(_debugIsConnected, '');
+    _logger.info('insertPendingOperation');
+    return db!.pendingOperationDao.insertPendingOperation(operation);
+  }
+
+  @override
+  Future<List<PendingOperation>> getPendingOperations() {
+    assert(_debugIsConnected, '');
+    _logger.info('getPendingOperations');
+    return db!.pendingOperationDao.getPendingOperations();
+  }
+
+  @override
+  Future<void> deletePendingOperation(int id) {
+    assert(_debugIsConnected, '');
+    _logger.info('deletePendingOperation');
+    return db!.pendingOperationDao.deletePendingOperation(id);
+  }
+
   bool _sortRequiresMembership(SortOrder<ChannelState>? sort) =>
       sort?.any((opt) => opt.field == ChannelSortKey.pinnedAt) ?? false;